send emails

+++ application/controllers/tasks.php
@@ -1,22 +1,22 @@
-<?php 
+<?php
 class Tasks extends CI_Controller {
 
 	function __construct() {
 		parent::__construct();
 		$this->load->model('Tasks_Model');
 	}
-	
+
 	function index() {
 	}
-	
+
 	function listing() {
 		if (! $this->safety->allowByControllerName(__METHOD__) ) { return errorForbidden(); }
-		
+
 		$page = (int)$this->input->get('page');
 		if ($page == 0) { $page = 1; }
 
 		$this->load->model('Tasks_Status_Model');
-		
+
 		$filters = array(
 			'search'       => $this->input->get('search'),
 			'taskRunning'  => ($this->input->get('taskRunning') === false ? null : $this->input->get('taskRunning')),
@@ -24,7 +24,7 @@ class Tasks extends CI_Controller {
 		$orders  = array(
 			array('orderBy' => $this->input->get('orderBy'), 'orderDir' => $this->input->get('orderDir') ),
 		);
-		
+
 		$query = $this->Tasks_Model->selectToList($page, config_item('pageSize'), $filters, $orders );
 
 		$this->load->view('pageHtml', array(
@@ -39,7 +39,7 @@ class Tasks extends CI_Controller {
 					'statusTaskName'    => $this->lang->line('Running'),
 					'langName'          => $this->lang->line('Language'),
 					'taskRetries'       => $this->lang->line('Retries'),
-					'taskSchedule'      => array('value' => $this->lang->line('Schedule date'), 'class' => 'datetime'), 
+					'taskSchedule'      => array('value' => $this->lang->line('Schedule date'), 'class' => 'datetime'),
 				),
 				'data'        => $query['data'],
 				'foundRows'   => $query['foundRows'],
@@ -60,15 +60,15 @@ class Tasks extends CI_Controller {
 				)
 			)
 		));
-	}	
-	
+	}
+
 	/*
 	 * Metodo que se llama desde un cronjobs para iniciar  el envio de las tasks_email
 	 */
 	function sendEmails(){
 		set_time_limit(0);
 		if(!$this->input->is_cli_request()){return error404();}
-		
+
 		switch (ENVIRONMENT) {
 			case 'development':
 				$this->config->set_item('base_url', config_item('urlDev'));
@@ -82,54 +82,53 @@ class Tasks extends CI_Controller {
 		}
 
 		$filters = array(
-			'taskRunning'  => TASK_PENDING,
+			'taskRunning'  => false,
+			'statusTaskId' => TASK_PENDING,
 			'validDate'    => true
-		);		
-		
+		);
+
 		$query   = $this->Tasks_Model->selectToList(1, 100, $filters, array() );
 		$rsTasks = $query['data'];
 		if(!empty($rsTasks)){
 			$this->load->library('SendMails');
-			
+
 			foreach ($rsTasks as $task) {
 				$task['taskRunning'] = TASK_RUNNING;
 				$this->Tasks_Model->save($task);
-				try {
-					//Sino se envio la tarea genero una excepcion
-					if(!$this->_sendEmail($task)){
-						throw new Exception('No se pode enviar el correo - taskId: '.$task['taskId']);
-					}
-				} catch (Exception $e) {
-					
+
+				$success = $this->_sendEmail($task);
+
+				if ($success == true) { //Cuando se completo el envio borro la tarea
+					$this->Tasks_Model->delete($task['taskId']);
+				}
+				else { //Sino se envio el email, aumento el contador de reintentos
 					if($task['taskRetries'] < TASK_RETRY){
 						//Cantidad de Reintentos
 						$task['taskRunning'] = TASK_PENDING;
 						$task['taskRetries'] = $task['taskRetries'] + 1;
 						$this->Tasks_Model->save($task);
-					}else{
+					}
+					else {
 						//Cambio el estado a Cancelado
 						$task['taskRunning'] = TASK_CANCEL;
 						$this->Tasks_Model->save($task);
 					}
-					continue;
 				}
-				//Cuando se completo el envio borro la tarea
-				$this->Tasks_Model->delete($task['taskId']);
 			}
 		}
 	}
-	
+
 	/*
-	 * 
+	 *
 	 * Metodo que ejecuta cada una de las tareas de envio de email, lo llama self::sendEmails
 	 * El array que recibe debe tener el indice taskMethod obligatorio
 	 * @param array task
 	 * @return boolean
-	 * 
-	 */	
+	 *
+	 */
 	function _sendEmail($task) {
 		if(!$this->input->is_cli_request()){ return error404(); }
-		
+
 		$return = false;
 		if(empty($task) || !is_array($task) || empty($task['taskMethod']) ){
 			return $return;
@@ -145,11 +144,10 @@ class Tasks extends CI_Controller {
 
 		if (method_exists($this->sendmails, $taskMethod)) {
 			if(!empty($taskParams)){
-				$this->sendmails->$taskMethod((array)json_decode($taskParams));
+				$return = $this->sendmails->$taskMethod((array)json_decode($taskParams));
 			}else{
-				$this->sendmails->$taskMethod();
+				$return = $this->sendmails->$taskMethod();
 			}
-			$return = true;
 		}
 
 		unset($taskMethod);
@@ -157,4 +155,3 @@ class Tasks extends CI_Controller {
 		return $return;
 	}
 }
-
